requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.23
Django==1.9.6
flake8==2.5.4
kombu==3.0.35
mccabe==0.4.0
pep8==1.7.0
psycopg2==2.6.1
pyflakes==1.0.0
pytz==2016.4
redis==2.10.3
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.23
Django==1.9.6
djangorestframework==3.3.3
flake8==2.5.4
kombu==3.0.35
mccabe==0.4.0
oauthlib==1.1.2
pep8==1.7.0
psycopg2==2.6.1
pyflakes==1.0.0
pytz==2016.4
redis==2.10.3
requests==2.10.0
requests-oauthlib==0.6.1
six==1.10.0
tweepy==3.5.0
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.23
Django==1.9.6
djangorestframework==3.3.3
flake8==2.5.4
kombu==3.0.35
mccabe==0.4.0
oauthlib==1.1.2
pep8==1.7.0
psycopg2==2.6.1
pyflakes==1.0.0
pygal==2.2.3
pytz==2016.4
redis==2.10.3
requests==2.10.0
requests-oauthlib==0.6.1
six==1.10.0
tweepy==3.5.0
